tests/pyproject.toml
[tool.poetry]
name = "databend"
version = "0.1.0"
description = "Apache-2.0"
authors = ["Databend Authors <opensource@datafuselabs.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.1"
pyre-check = "^0.9.10"
boto3 = "^1.20.38"
moto = {extras = ["all"], version = "^2.3.2"}
yapf = "^0.32.0"
shfmt-py = "^3.3.1"
mysql-connector-python = "^8.0.28"
SQLAlchemy = "^1.4.35"
clickhouse-driver = "^0.2.3"
PyMySQL = "^1.0.2"
six = "^1.16.0"
PyHamcrest = "^2.0.2"
requests = "^2.27.1"
environs = "^9.5.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
